Added an explicit definition for Show Player and made Player not derive Read for now.
The explicit declaration for Show should make the resulting String from Player be much more readable by humans.

Due to explicitly declaring Show and not knowing how to do the same for Read yet, I had to remove it for now, but I'll try to add it as soon as possible.