Reworked the getting indexes part of getValidMove.
Added support for exiting the game early by not writing any indexes and
wrote that in the starting message to inform players about it.

Also changed it so that it doesn't break/crash on getting only 1 index.

And shortened the code that gets both indexes, because it was
unnecessarily long.