Moved isGameWon to Grid.hs and renamed it to isGridWon.
Moved it because it makes much more sense in Grid.hs as it's just a
function that operates on a grid and doesn't do any IO or anything that
would warrant it being in Main.

Name was changed to better reflect what the function does even though it
still serves as a way to check if the game has been won.