Overhauled the type of playing Grid and created Either3.
Changed the type of PlayingGrid to more granual type of Grid, that can
represent when the grid is won, tied or still being used, so that it's
impossible for won/tied game/grid to be used and manipulated as if it was still
in use.

As a result the type signature of many functions has changed even though
the functions haven't actually changed.

Created Either3 that is basically Either a (Either b c), but shorter
and more readable.

Also created finalizeGridIfGameFinished (probably not the best name, but
at least it describes what it does) and used it in playGame instead of the
blob of ifs and elses to make things more readable and typesafe.